Rewrite Counting Heaps from CEPC 2008
I was thinking about the problem the other day, and completely forgot
that I had already solved it. The previous solution was not neither
readable nor generic, so I suppose this is a bit of an improvement. The
first part of the solution is a telescoping argument that is relatively
easy to come up with -- the expression we are trying to compute is

    (s(1) - 1)! / (s(2) * s(3) * ... * s(n))

where s(i) is the size of the subtree rooted at node i. (Note that in
this case, the root is always node 1.)

The tricky part in computing this (for me at least) is that the modulus
m is not always prime, so we cannot always easily compute the inverse of
s(i).

The way to do this is to note that we can actually factor all of the
numbers < n relatively quickly -- in O(n * log(log(n))), to be precise.
The very important lesson here is that you should not be afraid of
factoring lots of numbers, despite all the warnings you probably heard
in your first algorithms class! This is a lesson that I had somehow
forgotten since the time when I first solved this problem. The algorithm
and a proof of its runtime are included in the code and commit.